requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==36.0.1
defusedxml==0.7.1
distlib==0.3.3
Django==3.2
django-allauth==0.48.0
django-appconf==1.0.5
django-bootstrap-form==3.4
django-crispy-forms==1.14.0
django-debug-toolbar==3.6.0
django-etc==1.3.1
django-fancybox==0.1.6
django-filter==22.1
django-hitcount==1.3.5
django-material==1.10.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.4.0
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
pep8==1.7.1
Pillow==9.0.1
pinax-messages==3.0.0
pinax-templates==2.0.6
platformdirs==2.4.0
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.10.0
whitenoise==6.0.0
